feat: app navigation
