+++ components/common/buttons/index.ts
@@ -1,3 +1,4 @@
 import AppButton from "./AppButton";
+import BackButton from "./BackButton";
 
-export {AppButton}
+export {AppButton, BackButton}
