+++ components/common/inputs/index.ts
@@ -1,4 +1,5 @@
 import CustomInput from "./CustomInput";
 import CustomPhoneInput from "./CustomPhoneInput";
 
+
 export{CustomInput,CustomPhoneInput}
