+++ package.json
@@ -10,12 +10,16 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-navigation/bottom-tabs": "^6.5.7",
+    "@react-navigation/native": "^6.1.6",
+    "@react-navigation/native-stack": "^6.9.12",
     "nativewind": "^2.0.11",
     "react": "18.2.0",
     "react-native": "0.71.6",
     "react-native-paper": "^5.7.2",
     "react-native-phone-number-input": "^2.1.0",
-    "react-native-safe-area-context": "^4.5.2",
+    "react-native-safe-area-context": "^4.5.3",
+    "react-native-screens": "^3.20.0",
     "react-native-vector-icons": "^9.2.0"
   },
   "devDependencies": {
