+++ src/navigation/AuthNavigation.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+
+import { Register,Login } from '../screens/auth';
+
+export type RootStackParamList = {
+  Login: undefined;
+  Register: undefined;
+};
+
+const Stack = createNativeStackNavigator<RootStackParamList>();
+
+const AuthNavigation = () => {
+  return (
+    <Stack.Navigator
+      initialRouteName="Login"
+      screenOptions={{headerShown: false}}>
+      <Stack.Screen name="Login" component={Login} />
+      <Stack.Screen name="Register" component={Register} />
+    </Stack.Navigator>
+  );
+};
+
+export default AuthNavigation;
+
