+++ src/screens/auth/Register.tsx
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const Register = () => {
+  return (
+    <View>
+      <Text>Register</Text>
+    </View>
+  )
+}
+
+export default Register
+
+const styles = StyleSheet.create({})
