+++ src/screens/auth/index.tsx
@@ -1,0 +1,4 @@
+import Login from "./Login";
+import Register from "./Register";
+
+export {Login,Register}
